requirements.txt
requests
python-telegram-bot==13.11
loguru
chalice
boto3
wget
openai
langchain
pinecone-client
# pandas
# tqdm
# tiktoken

